lib/http: Move HTTP object serialization logic to lib/http
